Fix the coding style
