+++ src/test/java/seedu/address/logic/commands/redo/command/test/RedoClearCommandTest.java
@@ -1,14 +1,9 @@
-package seedu.address.logic.commands.RedoCommandTest;
+package seedu.address.logic.commands.redo.command.test;
 
 import static seedu.address.logic.UndoRedoStackUtil.prepareStack;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.address.logic.commands.CommandTestUtil.modifyPerson;
 import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
-import static seedu.address.testutil.TypicalPersons.CARL;
-import static seedu.address.testutil.TypicalPersons.DANIEL;
-import static seedu.address.testutil.TypicalPersons.IDA;
-import static seedu.address.testutil.TypicalPersons.HOON;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -18,40 +13,39 @@ import org.junit.Test;
 
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.UndoRedoStack;
-import seedu.address.logic.commands.EditCommand;
+import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.RedoCommand;
+import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 
-public class RedoEditCommandTest {
+public class RedoClearCommandTest {
     private static final CommandHistory EMPTY_COMMAND_HISTORY = new CommandHistory();
     private static final UndoRedoStack EMPTY_STACK = new UndoRedoStack();
 
     private final Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
-    private final EditCommand editCommandOne = new EditCommand(CARL, IDA);
-    private final EditCommand editCommandTwo = new EditCommand(DANIEL, HOON);
+    private final ClearCommand clearCommandOne = new ClearCommand();
+    private final ClearCommand clearCommandTwo = new ClearCommand();
 
     @Before
     public void setUp() {
-        editCommandOne.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
-        editCommandTwo.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
+        clearCommandOne.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
+        clearCommandTwo.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
     }
 
     @Test
     public void execute() {
         UndoRedoStack undoRedoStack = prepareStack(
-                Collections.emptyList(), Arrays.asList(editCommandTwo, editCommandOne));
+                Collections.emptyList(), Arrays.asList(clearCommandTwo, clearCommandOne));
         RedoCommand redoCommand = new RedoCommand();
         redoCommand.setData(model, EMPTY_COMMAND_HISTORY, undoRedoStack);
-        Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
+        Model expectedModel = new ModelManager(new AddressBook(), new UserPrefs());
 
         // multiple commands in redoStack
-        modifyPerson(expectedModel, CARL, IDA);
         assertCommandSuccess(redoCommand, model, RedoCommand.MESSAGE_SUCCESS, expectedModel);
 
         // single command in redoStack
-        modifyPerson(expectedModel, DANIEL, HOON);
         assertCommandSuccess(redoCommand, model, RedoCommand.MESSAGE_SUCCESS, expectedModel);
 
         // no command in redoStack
