+++ src/test/java/seedu/address/logic/commands/redo/command/test/RedoEditCommandTest.java
@@ -1,14 +1,14 @@
-package seedu.address.logic.commands.UndoCommandTest;
+package seedu.address.logic.commands.redo.command.test;
 
 import static seedu.address.logic.UndoRedoStackUtil.prepareStack;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.logic.commands.CommandTestUtil.modifyPerson;
-import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 import static seedu.address.testutil.TypicalPersons.CARL;
 import static seedu.address.testutil.TypicalPersons.DANIEL;
 import static seedu.address.testutil.TypicalPersons.IDA;
 import static seedu.address.testutil.TypicalPersons.HOON;
+import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -19,13 +19,12 @@ import org.junit.Test;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.UndoRedoStack;
 import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.UndoCommand;
+import seedu.address.logic.commands.RedoCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 
-
-public class UndoEditCommandTest {
+public class RedoEditCommandTest {
     private static final CommandHistory EMPTY_COMMAND_HISTORY = new CommandHistory();
     private static final UndoRedoStack EMPTY_STACK = new UndoRedoStack();
 
@@ -40,24 +39,22 @@ public class UndoEditCommandTest {
     }
 
     @Test
-    public void execute() throws Exception {
+    public void execute() {
         UndoRedoStack undoRedoStack = prepareStack(
-                Arrays.asList(editCommandOne, editCommandTwo), Collections.emptyList());
-        UndoCommand undoCommand = new UndoCommand();
-        undoCommand.setData(model, EMPTY_COMMAND_HISTORY, undoRedoStack);
+                Collections.emptyList(), Arrays.asList(editCommandTwo, editCommandOne));
+        RedoCommand redoCommand = new RedoCommand();
+        redoCommand.setData(model, EMPTY_COMMAND_HISTORY, undoRedoStack);
         Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
-        editCommandOne.execute();
-        editCommandTwo.execute();
 
-        // multiple commands in undoStack
+        // multiple commands in redoStack
         modifyPerson(expectedModel, CARL, IDA);
-        assertCommandSuccess(undoCommand, model, UndoCommand.MESSAGE_SUCCESS, expectedModel);
+        assertCommandSuccess(redoCommand, model, RedoCommand.MESSAGE_SUCCESS, expectedModel);
 
-        // single command in undoStack
-        expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
-        assertCommandSuccess(undoCommand, model, UndoCommand.MESSAGE_SUCCESS, expectedModel);
+        // single command in redoStack
+        modifyPerson(expectedModel, DANIEL, HOON);
+        assertCommandSuccess(redoCommand, model, RedoCommand.MESSAGE_SUCCESS, expectedModel);
 
-        // no command in undoStack
-        assertCommandFailure(undoCommand, model, UndoCommand.MESSAGE_FAILURE);
+        // no command in redoStack
+        assertCommandFailure(redoCommand, model, RedoCommand.MESSAGE_FAILURE);
     }
 }
