+++ src/test/java/seedu/address/logic/commands/undo/command/test/UndoClearCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.RedoCommandTest;
+package seedu.address.logic.commands.undo.command.test;
 
 import static seedu.address.logic.UndoRedoStackUtil.prepareStack;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
@@ -14,13 +14,14 @@ import org.junit.Test;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.UndoRedoStack;
 import seedu.address.logic.commands.ClearCommand;
-import seedu.address.logic.commands.RedoCommand;
+import seedu.address.logic.commands.UndoCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 
-public class RedoClearCommandTest {
+
+public class UndoClearCommandTest {
     private static final CommandHistory EMPTY_COMMAND_HISTORY = new CommandHistory();
     private static final UndoRedoStack EMPTY_STACK = new UndoRedoStack();
 
@@ -35,20 +36,24 @@ public class RedoClearCommandTest {
     }
 
     @Test
-    public void execute() {
+    public void execute() throws Exception {
         UndoRedoStack undoRedoStack = prepareStack(
-                Collections.emptyList(), Arrays.asList(clearCommandTwo, clearCommandOne));
-        RedoCommand redoCommand = new RedoCommand();
-        redoCommand.setData(model, EMPTY_COMMAND_HISTORY, undoRedoStack);
+                Arrays.asList(clearCommandOne, clearCommandTwo), Collections.emptyList());
+        UndoCommand undoCommand = new UndoCommand();
+        undoCommand.setData(model, EMPTY_COMMAND_HISTORY, undoRedoStack);
         Model expectedModel = new ModelManager(new AddressBook(), new UserPrefs());
+        clearCommandOne.execute();
+        clearCommandTwo.execute();
 
-        // multiple commands in redoStack
-        assertCommandSuccess(redoCommand, model, RedoCommand.MESSAGE_SUCCESS, expectedModel);
+        // multiple commands in undoStack
+        expectedModel.resetData(new AddressBook());
+        assertCommandSuccess(undoCommand, model, UndoCommand.MESSAGE_SUCCESS, expectedModel);
 
-        // single command in redoStack
-        assertCommandSuccess(redoCommand, model, RedoCommand.MESSAGE_SUCCESS, expectedModel);
+        // single command in undoStack
+        expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
+        assertCommandSuccess(undoCommand, model, UndoCommand.MESSAGE_SUCCESS, expectedModel);
 
-        // no command in redoStack
-        assertCommandFailure(redoCommand, model, RedoCommand.MESSAGE_FAILURE);
+        // no command in undoStack
+        assertCommandFailure(undoCommand, model, UndoCommand.MESSAGE_FAILURE);
     }
 }
