+++ src/test/java/seedu/address/logic/commands/undo/command/test/UndoDeleteCommandTest.java
@@ -1,12 +1,11 @@
-package seedu.address.logic.commands.UndoCommandTest;
+package seedu.address.logic.commands.undo.command.test;
 
 import static seedu.address.logic.UndoRedoStackUtil.prepareStack;
-import static seedu.address.logic.commands.CommandTestUtil.addPerson;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
+import static seedu.address.logic.commands.CommandTestUtil.deleteFirstPerson;
+import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
-import static seedu.address.testutil.TypicalPersons.HOON;
-import static seedu.address.testutil.TypicalPersons.IDA;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -16,38 +15,38 @@ import org.junit.Test;
 
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.UndoRedoStack;
-import seedu.address.logic.commands.AddCommand;
+import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.commands.UndoCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 
-public class UndoAddCommandTest {
+public class UndoDeleteCommandTest {
     private static final CommandHistory EMPTY_COMMAND_HISTORY = new CommandHistory();
     private static final UndoRedoStack EMPTY_STACK = new UndoRedoStack();
 
     private final Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
-    private final AddCommand addCommandOne = new AddCommand(HOON);
-    private final AddCommand addCommandTwo = new AddCommand(IDA);
+    private final DeleteCommand deleteCommandOne = new DeleteCommand(INDEX_FIRST_PERSON);
+    private final DeleteCommand deleteCommandTwo = new DeleteCommand(INDEX_FIRST_PERSON);
 
     @Before
     public void setUp() {
-        addCommandOne.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
-        addCommandTwo.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
+        deleteCommandOne.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
+        deleteCommandTwo.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
     }
 
     @Test
     public void execute() throws Exception {
         UndoRedoStack undoRedoStack = prepareStack(
-                Arrays.asList(addCommandOne, addCommandTwo), Collections.emptyList());
+                Arrays.asList(deleteCommandOne, deleteCommandTwo), Collections.emptyList());
         UndoCommand undoCommand = new UndoCommand();
         undoCommand.setData(model, EMPTY_COMMAND_HISTORY, undoRedoStack);
-        addCommandOne.execute();
-        addCommandTwo.execute();
+        deleteCommandOne.execute();
+        deleteCommandTwo.execute();
 
         // multiple commands in undoStack
         Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
-        addPerson(expectedModel, HOON);
+        deleteFirstPerson(expectedModel);
         assertCommandSuccess(undoCommand, model, UndoCommand.MESSAGE_SUCCESS, expectedModel);
 
         // single command in undoStack
