+++ src/test/java/seedu/address/logic/commands/undo/command/test/UndoEditCommandTest.java
@@ -1,9 +1,14 @@
-package seedu.address.logic.commands.UndoCommandTest;
+package seedu.address.logic.commands.undo.command.test;
 
 import static seedu.address.logic.UndoRedoStackUtil.prepareStack;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
+import static seedu.address.logic.commands.CommandTestUtil.modifyPerson;
+import static seedu.address.testutil.TypicalPersons.CARL;
+import static seedu.address.testutil.TypicalPersons.DANIEL;
+import static seedu.address.testutil.TypicalPersons.HOON;
 import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
+import static seedu.address.testutil.TypicalPersons.IDA;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -13,40 +18,39 @@ import org.junit.Test;
 
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.UndoRedoStack;
-import seedu.address.logic.commands.ClearCommand;
+import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.UndoCommand;
-import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 
 
-public class UndoClearCommandTest {
+public class UndoEditCommandTest {
     private static final CommandHistory EMPTY_COMMAND_HISTORY = new CommandHistory();
     private static final UndoRedoStack EMPTY_STACK = new UndoRedoStack();
 
     private final Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
-    private final ClearCommand clearCommandOne = new ClearCommand();
-    private final ClearCommand clearCommandTwo = new ClearCommand();
+    private final EditCommand editCommandOne = new EditCommand(CARL, IDA);
+    private final EditCommand editCommandTwo = new EditCommand(DANIEL, HOON);
 
     @Before
     public void setUp() {
-        clearCommandOne.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
-        clearCommandTwo.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
+        editCommandOne.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
+        editCommandTwo.setData(model, EMPTY_COMMAND_HISTORY, EMPTY_STACK);
     }
 
     @Test
     public void execute() throws Exception {
         UndoRedoStack undoRedoStack = prepareStack(
-                Arrays.asList(clearCommandOne, clearCommandTwo), Collections.emptyList());
+                Arrays.asList(editCommandOne, editCommandTwo), Collections.emptyList());
         UndoCommand undoCommand = new UndoCommand();
         undoCommand.setData(model, EMPTY_COMMAND_HISTORY, undoRedoStack);
-        Model expectedModel = new ModelManager(new AddressBook(), new UserPrefs());
-        clearCommandOne.execute();
-        clearCommandTwo.execute();
+        Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
+        editCommandOne.execute();
+        editCommandTwo.execute();
 
         // multiple commands in undoStack
-        expectedModel.resetData(new AddressBook());
+        modifyPerson(expectedModel, CARL, IDA);
         assertCommandSuccess(undoCommand, model, UndoCommand.MESSAGE_SUCCESS, expectedModel);
 
         // single command in undoStack
